+++ erl/erltest/test_receiver.go
@@ -6,7 +6,6 @@ package erltest
 
 import (
 	"errors"
-	"fmt"
 	"reflect"
 	"sync"
 	"testing"
@@ -21,23 +20,51 @@ import (
 	"github.com/uberbrodt/erl-go/erl/genserver"
 )
 
-var testTimeout time.Duration = chronos.Dur("10s")
+var (
+	DefaultReceiverTimeout time.Duration = chronos.Dur("10s")
+	DefaultWaitTimeout     time.Duration = chronos.Dur("5s")
+)
 
-type expectation struct {
-	h          TestExpectation
-	callHandle TestCallExpectation
-	opts       expectOpts
-	matchCnt   int
-	satisfied  bool
+type ReceiverOpt func(ro receiverOptions) receiverOptions
+
+type receiverOptions struct {
+	timeout     time.Duration
+	waitTimeout time.Duration
+}
+
+// Specify how long the test reciever should run for before stopping.
+// this needs to be set otherwise tests will hang until exceptions are matched or
+// the 10min Go default is reached. See [DefaultReceiverTimeout]
+func ReceiverTimeout(t time.Duration) ReceiverOpt {
+	return func(ro receiverOptions) receiverOptions {
+		ro.timeout = t
+		return ro
+	}
+}
+
+// Specify how long for [TestReciever.Wait] for all expectations to be met
+// see [DefaultWaitTimeout]
+func WaitTimeout(t time.Duration) ReceiverOpt {
+	return func(ro receiverOptions) receiverOptions {
+		ro.waitTimeout = t
+		return ro
+	}
 }
 
 // Creates a new TestReceiver, which is a process that you can set
 // message matching expectations on.
-func NewReceiver(t *testing.T) (erl.PID, *TestReceiver) {
-	expectations := make(map[reflect.Type]*expectation)
-	castExpects := make(map[reflect.Type]*expectation)
-	callExpects := make(map[reflect.Type]*expectation)
-	tr := &TestReceiver{t: t, expectations: expectations, castExpects: castExpects, callExpects: callExpects}
+func NewReceiver(t *testing.T, opts ...ReceiverOpt) (erl.PID, *TestReceiver) {
+	rOpts := receiverOptions{
+		timeout:     DefaultReceiverTimeout,
+		waitTimeout: DefaultWaitTimeout,
+	}
+	for _, o := range opts {
+		rOpts = o(rOpts)
+	}
+	expectations := make(map[reflect.Type]*Expectation)
+	castExpects := make(map[reflect.Type]*Expectation)
+	callExpects := make(map[reflect.Type]*Expectation)
+	tr := &TestReceiver{t: t, expectations: expectations, castExpects: castExpects, callExpects: callExpects, opts: rOpts}
 	pid := erl.Spawn(tr)
 
 	erl.ProcessFlag(pid, erl.TrapExit, true)
@@ -49,15 +76,17 @@ func NewReceiver(t *testing.T) (erl.PID, *TestReceiver) {
 
 type TestReceiver struct {
 	t            *testing.T
-	expectations map[reflect.Type]*expectation
-	castExpects  map[reflect.Type]*expectation
-	callExpects  map[reflect.Type]*expectation
+	expectations map[reflect.Type]*Expectation
+	castExpects  map[reflect.Type]*Expectation
+	callExpects  map[reflect.Type]*Expectation
 	failures     []*ExpectationFailure
 	msgCnt       int
 	self         erl.PID
 	// if set to true, t.FailNow will not be called in [Pass] or [Wait]
-	noFail bool
-	mx     sync.RWMutex
+	noFail    bool
+	mx        sync.RWMutex
+	testEnded bool
+	opts      receiverOptions
 }
 
 func (tr *TestReceiver) Receive(self erl.PID, inbox <-chan any) error {
@@ -85,7 +114,7 @@ func (tr *TestReceiver) Receive(self erl.PID, inbox <-chan any) error {
 			tr.mx.Lock()
 			tr.check(msg)
 			tr.mx.Unlock()
-		case <-time.After(testTimeout):
+		case <-time.After(tr.opts.timeout):
 			tr.t.Fatal("TestReceiver: test timeout")
 
 			return exitreason.Timeout
@@ -102,7 +131,7 @@ func (tr *TestReceiver) check(msg any) {
 		for match, ex := range tr.castExpects {
 			if castMsgT == match {
 				// pass in the unwrapped message
-				fail := tr.doCheck(match, v.Msg, ex)
+				fail := tr.checkMatch(match, v.Msg, ex)
 				if fail != nil {
 					tr.failures = append(tr.failures, fail)
 				}
@@ -117,7 +146,7 @@ func (tr *TestReceiver) check(msg any) {
 					return ex.callHandle(self, v.From, msg)
 				}
 				// pass in the unwrapped message
-				fail := tr.doCheck(match, v.Msg, ex)
+				fail := tr.checkMatch(match, v.Msg, ex)
 				if fail != nil {
 					tr.failures = append(tr.failures, fail)
 				}
@@ -127,7 +156,7 @@ func (tr *TestReceiver) check(msg any) {
 		msgT := reflect.TypeOf(msg)
 		for match, ex := range tr.expectations {
 			if msgT == match {
-				fail := tr.doCheck(match, msg, ex)
+				fail := tr.checkMatch(match, msg, ex)
 				if fail != nil {
 					tr.failures = append(tr.failures, fail)
 				}
@@ -137,119 +166,29 @@ func (tr *TestReceiver) check(msg any) {
 	}
 }
 
-func (tr *TestReceiver) doCheck(match reflect.Type, msg any, ex *expectation) (failure *ExpectationFailure) {
-	defer func() {
-		if r := recover(); r != nil {
-			failure = &ExpectationFailure{
-				MatchType: match,
-				Msg:       msg,
-				Reason:    fmt.Sprintf("TestExpectation panicked!: %+v", r),
-			}
-		}
-	}()
-	ex.matchCnt = ex.matchCnt + 1
-
-	// look for failures
-	switch ex.opts.exType {
-	case absolute:
-		// if we are matching absolute order, fail if the opts.times != the nth msg received
-		if tr.msgCnt != ex.opts.times {
-			return &ExpectationFailure{
-				MatchType: match,
-				Msg:       msg,
-				Reason:    fmt.Sprintf("expected to match msg #%d, but matched with msg #%d", ex.opts.times, tr.msgCnt),
-			}
-		}
-	case exact:
-		if ex.matchCnt > ex.opts.times {
-			ex.satisfied = false
-			return &ExpectationFailure{
-				MatchType: match,
-				Msg:       msg,
-				Reason:    fmt.Sprintf("expected to match %d times, but match count is now: %d", ex.opts.times, ex.matchCnt),
-			}
-		}
-	case atMost:
-		if ex.matchCnt > ex.opts.times {
-			return &ExpectationFailure{
-				MatchType: match,
-				Msg:       msg,
-				Reason:    fmt.Sprintf("expected to match at most %d times, but match count is now: %d", ex.opts.times, ex.matchCnt),
-			}
-		}
-	}
-
-	// if we're not matching any # of times and our matchCnt doesn't equal the times called, fail. This assumes
-	// we've already incremented the matchCnt
-
-	// if we returned false on a match, this is a failure.
-	if result := ex.h(tr.self, msg); !result {
-		return &ExpectationFailure{MatchType: match, Msg: msg, Reason: "returned false"}
-	}
-
-	// mark satisifed
-	switch ex.opts.exType {
-	case absolute:
-		if tr.msgCnt == ex.opts.times {
-			ex.satisfied = true
-		}
-	case exact:
-		ex.satisfied = ex.opts.times == ex.matchCnt
-	case atLeast:
-		ex.satisfied = ex.matchCnt >= ex.opts.times
-	}
-
-	return failure
+func (tr *TestReceiver) checkMatch(match reflect.Type, msg any, ex *Expectation) (failure *ExpectationFailure) {
+	arg := ExpectArg{Match: match, Msg: msg, Self: tr.self, MsgCount: tr.msgCnt}
+	return ex.Check(arg)
 }
 
 // Set an expectation that will be matched whenever an [expected] msg type is received.
 func (tr *TestReceiver) Expect(expected any, handler TestExpectation, opts ...ExpectOpt) {
-	o := expectOpts{times: 1, exType: exact}
-
-	for _, f := range opts {
-		o = f(o)
-	}
-
 	t := reflect.TypeOf(expected)
-	tr.expectations[t] = &expectation{
-		h:         handler,
-		opts:      o,
-		satisfied: o.exType == anyTimes || o.exType == atMost,
-	}
+	tr.expectations[t] = NewExpectation(handler, opts...)
 }
 
 // This is like [Expect] but is only tested against [genserver.CastRequest] messages.
 func (tr *TestReceiver) ExpectCast(expected any, handler TestExpectation, opts ...ExpectOpt) {
-	o := expectOpts{times: 1, exType: exact}
-
-	for _, f := range opts {
-		o = f(o)
-	}
-
 	t := reflect.TypeOf(expected)
-	tr.castExpects[t] = &expectation{
-		h:         handler,
-		opts:      o,
-		satisfied: o.exType == anyTimes || o.exType == atMost,
-	}
+	tr.castExpects[t] = NewExpectation(handler, opts...)
 }
 
 // This is like [Expect] but is only tested against [genserver.CallRequest] messages.
 // NOTE: You should use [genserver.Reply] to send a response to the [genserver.From], otherwise
 // the caller will timeout
 func (tr *TestReceiver) ExpectCall(expected any, handler TestCallExpectation, opts ...ExpectOpt) {
-	o := expectOpts{times: 1, exType: exact}
-
-	for _, f := range opts {
-		o = f(o)
-	}
-
 	t := reflect.TypeOf(expected)
-	tr.callExpects[t] = &expectation{
-		callHandle: handler,
-		opts:       o,
-		satisfied:  o.exType == anyTimes || o.exType == atMost,
-	}
+	tr.callExpects[t] = NewCallExpectation(handler, opts...)
 }
 
 // returns the number of failed expectations and whether
@@ -261,14 +200,14 @@ func (tr *TestReceiver) Pass() (int, bool) {
 	expects := maps.Values(tr.expectations)
 	castExpects := maps.Values(tr.castExpects)
 	callExpects := maps.Values(tr.callExpects)
-	reducer := func(list []*expectation) bool {
-		return fun.Reduce(list, true, func(v *expectation, acc bool) bool {
+	reducer := func(list []*Expectation) bool {
+		return fun.Reduce(list, true, func(v *Expectation, acc bool) bool {
 			tr.t.Logf("checking %s times: %d matchCnt: %d is satisifed: %t", v.opts.exType, v.opts.times, v.matchCnt, v.satisfied)
 			if !acc {
 				return acc
 			}
 
-			return v.satisfied
+			return v.Satisified(tr.testEnded)
 		})
 	}
 
@@ -276,46 +215,48 @@ func (tr *TestReceiver) Pass() (int, bool) {
 	return len(tr.failures), reducer(expects) && reducer(castExpects) && reducer(callExpects)
 }
 
-// Returns when the [tout] expires or an expectation fails. If at the end of the timeout not
+func (tr *TestReceiver) finish() bool {
+	fails, passed := tr.Pass()
+	if fails > 0 {
+		tr.mx.Lock()
+		defer tr.mx.Unlock()
+		tr.t.Logf("TestReceiver %s, had %d expectation failures\r", tr.self, fails)
+		for i, f := range tr.failures {
+			tr.t.Logf("Failure %d: [Reason: %s] [MatchType: %+v] [Msg: %+v]", i, f.Reason, f.MatchType, f.Msg)
+		}
+		// tr.t.Logf("Failures: %+v", tr.failures)
+		if !tr.noFail {
+			tr.t.FailNow()
+		} else {
+			return false
+		}
+	}
+	return passed
+}
+
+// Returns when the [WaitTimeout] expires or an expectation fails. If at the end of the timeout not
 // all expectations are satisifed, the test is failed. Call this after you have sent your messages
 // and want fail if your expecations don't pass
-func (tr *TestReceiver) WaitFor(tout time.Duration) {
+func (tr *TestReceiver) Wait() {
 	now := time.Now()
 	for {
-		if time.Since(now) > tout {
-			if !tr.noFail {
+		if time.Since(now) > tr.opts.waitTimeout {
+			tr.t.Logf("test ended, checking expectations a final time")
+			tr.testEnded = true
+			passed := tr.finish()
+			if !tr.noFail && !passed {
 				tr.t.Fatal("TestReceiver.Loop test timeout")
-			} else {
-				return
 			}
+			return
 		}
 
-		fails, passed := tr.Pass()
-		if fails > 0 {
-			tr.mx.Lock()
-			defer tr.mx.Unlock()
-			tr.t.Logf("TestReceiver %s, had %d expectation failures\r", tr.self, fails)
-			for i, f := range tr.failures {
-				tr.t.Logf("Failure %d: [Reason: %s] [MatchType: %+v] [Msg: %+v]", i, f.Reason, f.MatchType, f.Msg)
-			}
-			// tr.t.Logf("Failures: %+v", tr.failures)
-			if !tr.noFail {
-				tr.t.FailNow()
-			} else {
-				return
-			}
-		}
 		// no failures and all passed, return so test ends
+		passed := tr.finish()
 		if passed {
 			return
 		}
 		time.Sleep(time.Millisecond)
 	}
-}
-
-// same as [WaitFor], but uses default test timeout
-func (tr *TestReceiver) Wait() {
-	tr.WaitFor(testTimeout)
 }
 
 // will cause the test receiver to exit, sending signals to linked and monitoring